refactor : isOpen state hook화
