+++ src/components/modal/contents/CreateCategoryModal.tsx
@@ -1,0 +1,3 @@
+export default function CreateCategoryModal() {
+  return <div>CreateCategoryModal</div>;
+}
